+++ common/src/main/java/me/juancarloscp52/spyglass_improvements/integrations/AccessoriesIntegration.java
@@ -1,6 +1,6 @@
 package me.juancarloscp52.spyglass_improvements.integrations;
 
-import io.wispforest.accessories.api.*;
+//import io.wispforest.accessories.api.*;
 import me.juancarloscp52.spyglass_improvements.client.integrations.IEquipmentIntegration;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Items;
@@ -10,6 +10,7 @@ public class AccessoriesIntegration implements IEquipmentIntegration {
 
     @Override
     public boolean isPlayerUsingSpyglass(Player player) {
-        return AccessoriesCapability.get(player).isEquipped(Items.SPYGLASS);
+        return false;
+        //return AccessoriesCapability.get(player).isEquipped(Items.SPYGLASS);
     }
     }