+++ changelog.md
@@ -9,6 +9,15 @@ Find your version by looking for the file for your mod-loader:
 -   spyglass_improvements-`<mod-version>`+mc`<mc-version>`+forge.jar
 -   spyglass-improvements-`<mod-version>`+mc`<mc-version>`+fabric.jar
 
+## [1.5.12-beta for 1.21.11] - 2025-12-24 (fabric, neoforge)
+Available versions:
+-   Fabric: 1.21.11
+-   Neoforge: 1.21.11
+
+### Changes
+-   Port to 1.21.11
+-   Accessories mod has been disabled until it is updated.
+
 ## [1.5.12] - 2025-11-02 (fabric, neoforge, forge)
 Available versions:
 -   Fabric: 1.20, 1.20.1, 1.21.9, 1.21.10